Playing with valgrind and hexen2.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/branches/wip@3730 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/server/pr_cmds.c
@@ -3767,7 +3767,7 @@ static void QCBUILTIN PF_lightstylevalue (progfuncs_t *prinst, struct globalvars
 static void QCBUILTIN PF_lightstylestatic (progfuncs_t *prinst, struct globalvars_s *pr_globals)
 {
 	int		style;
-	float	num;
+	int	num;
 	char	*val;
 
 	static char *styleDefs[] =
@@ -3795,7 +3795,7 @@ static void QCBUILTIN PF_lightstylestatic (progfuncs_t *prinst, struct globalvar
 		num = 0;
 	else if (num >= 'z'-'a')
 		num = 'z'-'a'-1;
-	val = styleDefs[(int)num];
+	val = styleDefs[num];
 
 	PF_applylightstyle(style, val, col);
 }
